Carve `TestNetworkFixtures` out of the monolith (wala/ML#635 Phase 3).
The 31 end-to-end network fixtures (tutorials, LSTM/Transformer blocks,
the GNN family, CRF metrics) move verbatim; the monolith keeps 247,
preserving the module total.
